backend/requirements.txt
numpy<2
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
yt-dlp==2024.10.07
openai-whisper==20231117
vosk==0.3.45
transformers==4.44.2
torch==2.2.2
sentencepiece==0.2.0
accelerate==0.34.2
fpdf2==2.7.9
python-multipart==0.0.9
httpx==0.27.2
huggingface_hub==0.24.6
